requirements.txt
argon2-cffi==20.1.0
cffi==1.14.5
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
pymongo==3.11.3
six==1.15.0
uuid==1.30
Werkzeug==1.0.1
WTForms==2.3.3
